refactor(peagen): move task schemas
