+++ pkgs/standards/peagen/README.md
@@ -404,11 +404,11 @@ result, idx = pea.process_single_project(projects[0], start_idx=0)
 ### Transport Models
 
 Runtime RPC payloads should be validated using the Pydantic schemas generated
-in `peagen.models.schemas`. For example, use
+in `peagen.schemas`. For example, use
 `TaskRead.model_validate_json()` when decoding a task received over the network:
 
 ```python
-from peagen.models.schemas import TaskRead
+from peagen.schemas import TaskRead
 
 task = TaskRead.model_validate_json(raw_json)
 ```
